chore: improve the ownership model of the dialog
