+++ frontend/src/component/feature/FeatureStrategy/FeatureStrategyMenu/FeatureStrategyMenu.tsx
@@ -13,6 +13,9 @@ import { usePlausibleTracker } from 'hooks/usePlausibleTracker';
 import { useUiFlag } from 'hooks/useUiFlag';
 import { ReleasePlanAddChangeRequestDialog } from 'component/feature/FeatureView/FeatureOverview/ReleasePlan/ChangeRequest/ReleasePlanAddChangeRequestDialog';
 import type { IReleasePlanTemplate } from 'interfaces/releasePlans';
+import { useChangeRequestApi } from 'hooks/api/actions/useChangeRequestApi/useChangeRequestApi';
+import { usePendingChangeRequests } from 'hooks/api/getters/usePendingChangeRequests/usePendingChangeRequests';
+import useToast from 'hooks/useToast';
 
 interface IFeatureStrategyMenuProps {
     label: string;
@@ -57,6 +60,10 @@ export const FeatureStrategyMenu = ({
     const isPopoverOpen = Boolean(anchor);
     const popoverId = isPopoverOpen ? 'FeatureStrategyMenuPopover' : undefined;
     const flagOverviewRedesignEnabled = useUiFlag('flagOverviewRedesign');
+    const { setToastData } = useToast();
+    const { addChange } = useChangeRequestApi();
+    const { refetch: refetchChangeRequests } =
+        usePendingChangeRequests(projectId);
 
     const onClose = () => {
         setAnchor(undefined);
@@ -73,6 +80,25 @@ export const FeatureStrategyMenu = ({
 
     const openMoreStrategies = (event: React.SyntheticEvent) => {
         setAnchor(event.currentTarget);
+    };
+
+    const addReleasePlanToChangeRequest = async () => {
+        addChange(projectId, environmentId, {
+            feature: featureId,
+            action: 'addReleasePlan',
+            payload: {
+                templateId: templateForChangeRequestDialog?.id,
+            },
+        });
+
+        refetchChangeRequests();
+
+        setToastData({
+            type: 'success',
+            text: 'Added to draft',
+        });
+
+        setTemplateForChangeRequestDialog(undefined);
     };
 
     const createStrategyPath = formatCreateStrategyPath(
@@ -188,8 +214,9 @@ export const FeatureStrategyMenu = ({
                 />
             </Popover>
             <ReleasePlanAddChangeRequestDialog
-                projectId={projectId}
+                onConfirm={addReleasePlanToChangeRequest}
                 onClosing={() => setTemplateForChangeRequestDialog(undefined)}
+                isOpen={Boolean(templateForChangeRequestDialog)}
                 featureId={featureId}
                 environmentId={environmentId}
                 releaseTemplate={templateForChangeRequestDialog}
